templating, buffer, safe output
create templates for layout + main + off;
add data-file, config, functions,
safe output with htmlspecialchars;

correct style + alternative syntax (func)
