+++ config.php
@@ -1,0 +1,8 @@
+<?php
+
+$config = [
+    'sitename' => 'Дела в порядке',
+    'tpl_path' => 'templates/',
+    'enable' => true
+];
+?>
