+++ data.php
@@ -1,0 +1,43 @@
+<?php
+
+$projects = ["Все", "Входящие", "Учеба", "Работа", "Домашние дела", "Авто"];
+$tasks = [
+    [
+        'title' => 'Собеседование в IT компании',
+        'deadline' => '01.06.2018',
+        'category' => 'Работа',
+        'readiness' => false
+    ],
+    [
+        'title' => 'Выполнить тестовое задание',
+        'deadline' => '25.05.2018',
+        'category' => 'Работа',
+        'readiness' => false
+    ],
+    [
+        'title' => 'Сделать задание первого раздела',
+        'deadline' => '21.04.2018',
+        'category' => 'Учеба',
+        'readiness' => true
+    ],
+    [
+        'title' => 'Встреча с другом',
+        'deadline' => '	22.04.2018',
+        'category' => 'Входящие',
+        'readiness' => false
+    ],
+    [
+        'title' => 'Купить корм для кота',
+        'deadline' => 'Нет',
+        'category' => 'Домашние дела',
+        'readiness' => false
+    ],
+    [
+        'title' => 'Заказать пиццу',
+        'deadline' => 'Нет',
+        'category' => 'Домашние дела',
+        'readiness' => false
+    ]
+];
+
+?>
